initial breakdown of the extension.ts
