+++ packages/zowe-explorer/src/extension.ts
@@ -14,7 +14,6 @@ import * as globals from "./globals";
 import * as vscode from "vscode";
 import * as ussActions from "./uss/actions";
 import * as dsActions from "./dataset/actions";
-import * as jobActions from "./job/actions";
 import * as refreshActions from "./shared/refresh";
 import * as sharedActions from "./shared/actions";
 import {
@@ -29,9 +28,6 @@ import { ZoweExplorerApiRegister } from "./ZoweExplorerApiRegister";
 import { ZoweExplorerExtender } from "./ZoweExplorerExtender";
 import { Profiles } from "./Profiles";
 import { initializeZoweFolder, promptCredentials, readConfigFromDisk, writeOverridesFile } from "./utils/ProfilesUtils";
-import { createDatasetTree } from "./dataset/DatasetTree";
-import { createJobsTree } from "./job/ZosJobsProvider";
-import { createUSSTree } from "./uss/USSTree";
 import { MvsCommandHandler } from "./command/MvsCommandHandler";
 import SpoolProvider from "./SpoolProvider";
 import * as nls from "vscode-nls";
@@ -39,10 +35,9 @@ import { TsoCommandHandler } from "./command/TsoCommandHandler";
 import { cleanTempDir, moveTempFolder, hideTempFolder } from "./utils/TempFolder";
 import { SettingsConfig } from "./utils/SettingsConfig";
 import { handleSaving } from "./utils/workspace";
-import { ZoweDatasetNode } from "./dataset/ZoweDatasetNode";
-import * as contextuals from "../src/shared/context";
-import { Job } from "./job/ZoweJobNode";
-import { getSelectedNodeList } from "./shared/utils";
+import { initDatasetProvider } from "./dataset/extension";
+import { initUSSProvider } from "./uss/extension";
+import { initJobsProvider } from "./job/extension";
 
 // Set up localization
 nls.config({
@@ -105,13 +100,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         fs.mkdirSync(globals.DS_DIR);
     }
 
-    // Initialize dataset provider
-    datasetProvider = await createDatasetTree(globals.LOG);
-    // Initialize uss provider
-    ussFileProvider = await createUSSTree(globals.LOG);
-    // Initialize Jobs provider with the created session and the selected pattern
-    jobsProvider = await createJobsTree(globals.LOG);
-
     // set a command to silently reload extension
     context.subscriptions.push(
         vscode.commands.registerCommand("zowe.extRefresh", async () => {
@@ -150,6 +138,13 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
     );
     context.subscriptions.push(spoolProvider, providerRegistration);
 
+    // Initialize dataset provider
+    datasetProvider = await initDatasetProvider(context);
+    // Initialize uss provider
+    ussFileProvider = await initUSSProvider(context);
+    // Initialize Jobs provider with the created session and the selected pattern
+    jobsProvider = await initJobsProvider(context);
+
     // Register functions & event listeners
     context.subscriptions.push(
         vscode.workspace.onDidChangeConfiguration(async (e) => {
@@ -178,15 +173,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<ZoweEx
         })
     );
 
-    if (datasetProvider) {
-        initDatasetProvider(context);
-    }
-    if (ussFileProvider) {
-        initUSSProvider(context);
-    }
-    if (jobsProvider) {
-        initJobsProvider(context);
-    }
     if (datasetProvider || ussFileProvider) {
         context.subscriptions.push(
             vscode.commands.registerCommand("zowe.openRecentMember", () =>
@@ -339,554 +325,6 @@ async function watchConfigProfile(context: vscode.ExtensionContext) {
         workspaceProfileWatcher.onDidDelete(async () => {
             await vscode.commands.executeCommand("zowe.extRefresh");
         });
-}
-
-function initDatasetProvider(context: vscode.ExtensionContext) {
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.all.config.init", async () => {
-            datasetProvider.createZoweSchema(datasetProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.addSession", async () =>
-            datasetProvider.createZoweSession(datasetProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.addFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList);
-            for (const item of selectedNodes) {
-                datasetProvider.addFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.refreshAll", async () => {
-            await refreshActions.refreshAll(datasetProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.refreshNode", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isDsMember(element)
-            );
-            for (const item of selectedNodes) {
-                await dsActions.refreshPS(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.refreshDataset", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isPdsNotFav(element)
-            );
-            for (const item of selectedNodes) {
-                await dsActions.refreshDataset(item, datasetProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.pattern", (node) => datasetProvider.filterPrompt(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.editSession", async (node) =>
-            datasetProvider.editSession(node, datasetProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.ZoweNode.openPS", (node) =>
-            dsActions.openPS(node, true, datasetProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.createDataset", (node) => dsActions.createFile(node, datasetProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.createMember", (node) => dsActions.createMember(node, datasetProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.deleteDataset", (node?) =>
-            dsActions.deleteDatasetPrompt(datasetProvider, node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.allocateLike", (node) => dsActions.allocateLike(datasetProvider, node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.uploadDialog", (node) => dsActions.uploadDialog(node, datasetProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.deleteMember", (node?) =>
-            dsActions.deleteDatasetPrompt(datasetProvider, node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.editDataSet", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isDsMember(element)
-            );
-            for (const item of selectedNodes) {
-                await dsActions.openPS(item, false, datasetProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.editMember", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isDsMember(element)
-            );
-            for (const item of selectedNodes) {
-                await dsActions.openPS(item, false, datasetProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand(
-            "zowe.ds.removeSession",
-            async (node: IZoweDatasetTreeNode, nodeList: IZoweDatasetTreeNode[]) => {
-                let selectedNodes = getSelectedNodeList(node, nodeList);
-                selectedNodes = selectedNodes.filter((sNode) => contextuals.isDsSession(sNode));
-                for (const select of selectedNodes) {
-                    datasetProvider.deleteSession(select);
-                }
-            }
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.removeFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList);
-            for (const item of selectedNodes) {
-                await datasetProvider.removeFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.saveSearch", async (node) => datasetProvider.addFavorite(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.removeSavedSearch", async (node) =>
-            datasetProvider.removeFavorite(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.removeFavProfile", async (node) =>
-            datasetProvider.removeFavProfile(node.label, true)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.submitJcl", async () => dsActions.submitJcl(datasetProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.submitMember", async (node) => dsActions.submitMember(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.showAttributes", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList).filter(
-                (element) => contextuals.isDs(element) || contextuals.isPds(element) || contextuals.isDsMember(element)
-            );
-            for (const item of selectedNodes) {
-                await dsActions.showAttributes(item, datasetProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.renameDataSet", (node) => datasetProvider.rename(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.copyMember", (node) => dsActions.copyDataSet(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.copyDataSet", (node) => dsActions.copyDataSet(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.pasteMember", (node) => dsActions.pasteMember(node, datasetProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.renameDataSetMember", (node) => datasetProvider.rename(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.hMigrateDataSet", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter(
-                (element) => contextuals.isDs(element) || contextuals.isPdsNotFav(element)
-            );
-            for (const item of selectedNodes) {
-                dsActions.hMigrateDataSet(item as ZoweDatasetNode);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.hRecallDataSet", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter((element) => contextuals.isMigrated(element));
-            for (const item of selectedNodes) {
-                dsActions.hRecallDataSet(item as ZoweDatasetNode);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.showImperativeErrorDetails", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter((element) => contextuals.hasImperativeError(element));
-            for (const item of selectedNodes) {
-                dsActions.showImperativeErrorDetails(item as ZoweDatasetNode);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.disableValidation", async (node) =>
-            Profiles.getInstance().disableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.enableValidation", async (node) =>
-            Profiles.getInstance().enableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.ssoLogin", async (node: IZoweTreeNode) =>
-            datasetProvider.ssoLogin(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.ds.ssoLogout", async (node: IZoweTreeNode) =>
-            datasetProvider.ssoLogout(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.workspace.onDidChangeConfiguration((e) => {
-            datasetProvider.onDidChangeConfiguration(e);
-        })
-    );
-
-    initSubscribers(context, datasetProvider);
-}
-
-function initUSSProvider(context: vscode.ExtensionContext) {
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.addFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList);
-            for (const item of selectedNodes) {
-                await ussFileProvider.addFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList);
-            for (const item of selectedNodes) {
-                await ussFileProvider.removeFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.addSession", async () =>
-            ussFileProvider.createZoweSession(ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.refreshAll", async () => {
-            await refreshActions.refreshAll(ussFileProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.refreshUSS", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isDocument(x));
-            for (const item of selectedNodes) {
-                await item.refreshUSS();
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.refreshUSSInTree", (node: IZoweUSSTreeNode) =>
-            ussActions.refreshUSSInTree(node, ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.refreshDirectory", (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isUssDirectory(x));
-            for (const item of selectedNodes) {
-                ussActions.refreshDirectory(item, ussFileProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.fullPath", (node: IZoweUSSTreeNode) =>
-            ussFileProvider.filterPrompt(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.editSession", async (node) =>
-            ussFileProvider.editSession(node, ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.ZoweUSSNode.open", (node: IZoweUSSTreeNode) =>
-            node.openUSS(false, true, ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeSession", async (node: IZoweUSSTreeNode, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList);
-            selectedNodes = selectedNodes.filter((element) => contextuals.isUssSession(element));
-            for (const item of selectedNodes) {
-                ussFileProvider.deleteSession(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.createFile", async (node: IZoweUSSTreeNode) =>
-            ussActions.createUSSNode(node, ussFileProvider, "file")
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.createFolder", async (node: IZoweUSSTreeNode) =>
-            ussActions.createUSSNode(node, ussFileProvider, "directory")
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.deleteNode", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isDocument(x) || contextuals.isUssDirectory(x));
-            const cancelled = await ussActions.deleteUSSFilesPrompt(selectedNodes);
-            for (const item of selectedNodes) {
-                await item.deleteUSSNode(ussFileProvider, item.getUSSDocumentFilePath(), cancelled);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.binary", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isText(x));
-            for (const item of selectedNodes) {
-                await ussActions.changeFileType(item, true, ussFileProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.text", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweUSSTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isBinary(x));
-            for (const item of selectedNodes) {
-                await ussActions.changeFileType(item, false, ussFileProvider);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.renameNode", async (node: IZoweUSSTreeNode) =>
-            ussFileProvider.rename(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.uploadDialog", async (node: IZoweUSSTreeNode) =>
-            ussActions.uploadDialog(node, ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.copyPath", async (node: IZoweUSSTreeNode) =>
-            ussActions.copyPath(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.editFile", (node: IZoweUSSTreeNode) =>
-            node.openUSS(false, false, ussFileProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.saveSearch", async (node: IZoweUSSTreeNode) =>
-            ussFileProvider.saveSearch(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeSavedSearch", async (node: IZoweUSSTreeNode) =>
-            ussFileProvider.removeFavorite(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.removeFavProfile", async (node) =>
-            ussFileProvider.removeFavProfile(node.label, true)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.disableValidation", async (node) =>
-            Profiles.getInstance().disableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.enableValidation", async (node) =>
-            Profiles.getInstance().enableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.ssoLogin", async (node: IZoweTreeNode) =>
-            ussFileProvider.ssoLogin(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.uss.ssoLogout", async (node: IZoweTreeNode) =>
-            ussFileProvider.ssoLogout(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.workspace.onDidChangeConfiguration((e) => {
-            ussFileProvider.onDidChangeConfiguration(e);
-        })
-    );
-
-    initSubscribers(context, ussFileProvider);
-}
-
-function initJobsProvider(context: vscode.ExtensionContext) {
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.zosJobsOpenspool", (session, spool, refreshTimestamp) =>
-            jobActions.getSpoolContent(session, spool, refreshTimestamp)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.deleteJob", async (job, jobs) => {
-            await jobActions.deleteCommand(jobsProvider, job, jobs);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.runModifyCommand", (job) => jobActions.modifyCommand(job))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.runStopCommand", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList) as IZoweJobTreeNode[];
-            for (const item of selectedNodes) {
-                await jobActions.stopCommand(item as Job);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.refreshJobsServer", async (job) =>
-            jobActions.refreshJobsServer(job, jobsProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.refreshAllJobs", async () => {
-            await refreshActions.refreshAll(jobsProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.refreshJob", async (job) => {
-            jobActions.refreshJob(job.mParent, jobsProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.refreshSpool", async (node) => {
-            await jobActions.getSpoolContentFromMainframe(node);
-            jobActions.refreshJob(node.mParent.mParent, jobsProvider);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.addJobsSession", () => jobsProvider.createZoweSession(jobsProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.setOwner", (job) => jobActions.setOwner(job, jobsProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.setPrefix", (job) => jobActions.setPrefix(job, jobsProvider))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.removeJobsSession", (job, jobList) => {
-            let selectedNodes = getSelectedNodeList(job, jobList);
-            selectedNodes = selectedNodes.filter((element) => contextuals.isJobsSession(element));
-            for (const item of selectedNodes) {
-                jobsProvider.deleteSession(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.downloadSpool", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList) as IZoweJobTreeNode[];
-            await jobActions.downloadSpool(selectedNodes);
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.getJobJcl", async (node, nodeList) => {
-            let selectedNodes = getSelectedNodeList(node, nodeList) as IZoweJobTreeNode[];
-            selectedNodes = selectedNodes.filter((x) => contextuals.isJob(x));
-            for (const job of selectedNodes) {
-                await jobActions.downloadJcl(job as Job);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.setJobSpool", async (session, jobId) =>
-            jobActions.focusOnJob(jobsProvider, session, jobId)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.search", (node) => jobsProvider.filterPrompt(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.editSession", async (node) =>
-            jobsProvider.editSession(node, jobsProvider)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.addFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList) as IZoweJobTreeNode[];
-            for (const item of selectedNodes) {
-                await jobsProvider.addFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.removeFavorite", async (node, nodeList) => {
-            const selectedNodes = getSelectedNodeList(node, nodeList) as IZoweJobTreeNode[];
-            for (const item of selectedNodes) {
-                await jobsProvider.removeFavorite(item);
-            }
-        })
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.saveSearch", async (node) => jobsProvider.saveSearch(node))
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.removeSearchFavorite", async (node) =>
-            jobsProvider.removeFavorite(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.removeFavProfile", async (node) =>
-            jobsProvider.removeFavProfile(node.label, true)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.disableValidation", async (node) =>
-            Profiles.getInstance().disableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.enableValidation", async (node) =>
-            Profiles.getInstance().enableValidation(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.ssoLogin", async (node: IZoweTreeNode) =>
-            jobsProvider.ssoLogin(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.commands.registerCommand("zowe.jobs.ssoLogout", async (node: IZoweTreeNode) =>
-            jobsProvider.ssoLogout(node)
-        )
-    );
-    context.subscriptions.push(
-        vscode.workspace.onDidChangeConfiguration((e) => {
-            jobsProvider.onDidChangeConfiguration(e);
-        })
-    );
-
-    initSubscribers(context, jobsProvider);
 }
 
 function initSubscribers(context: vscode.ExtensionContext, theProvider: IZoweTree<IZoweTreeNode>) {
